feat: add GalleryFragment and GalleryViewModel classes
The GalleryFragment class is added to the project. It extends the Fragment class and is responsible for displaying the gallery UI. It inflates the FragmentGalleryBinding layout and sets up the ViewModel to observe changes in the text value. The onDestroyView method is overridden to clean up the binding.

The GalleryViewModel class is added to the project. It extends the ViewModel class and provides a LiveData object for the text value. The initial value of the text is set to "This is gallery Fragment".