feat(slideshow): add SlideshowFragment and SlideshowViewModel classes
The SlideshowFragment class is responsible for displaying a slideshow in the app. It extends the Fragment class and overrides the onCreateView() method to inflate the layout and set up the necessary views. It also uses a SlideshowViewModel to retrieve and display the slideshow content.

The SlideshowViewModel class is a ViewModel that holds the data for the SlideshowFragment. It has a LiveData<String> property called text, which represents the content of the slideshow. The initial value of the text is set to "This is slideshow Fragment".

These classes are added to the app/src/main/java/co/edu/uniandes/vinilos/ui/slideshow package.